Scoring
Gestion du calcul du score, ainsi que de son affichage dans l'HUD

+++ Pang Remake/Assets/Scripts/AllEvents.cs
@@ -52,6 +52,6 @@ public class QuitButtonClickedEvent : SDD.Events.Event
 #region Score Event
 public class ScoreItemEvent : SDD.Events.Event
 {
-	public int eScore;
+	public float eScore;
 }
 #endregion
